% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_sampling}
\alias{add_sampling}
\title{add_sampling}
\usage{
add_sampling(model)
}
\arguments{
\item{model}{A paratb model object}
}
\value{
A modified model object
}
\description{
add_sampling
}
\details{
Add a couple of columns to the select matrix to do blood and milk
sampling. This will require scheduling birth events with select 9
and 10
}
\author{
Thomas Rosendal
}
