% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_surveillance_event}
\alias{add_surveillance_event}
\title{add_surveillance_events}
\usage{
add_surveillance_event(
  model,
  i,
  t = 1,
  n = 1,
  events = as(model@events, "data.frame")
)
}
\arguments{
\item{model}{The model object}

\item{i}{the node to add a surveillance event to}

\item{t}{the timepoint that you want to do surveillance}

\item{n}{the number of animals to seed}

\item{events}{A dataframe of the events the default is just to
extract them from the model object.}
}
\value{
a paratb model object with some extra events
}
\description{
Add surveillance events
}
\details{
A method for adding surveillance events to the model. These are
events of Event type == 1 that increment the BL or BM compartments
at a specific time point. These require select = 9 for blood
sampling of adults and select = 10 for bulk milk sampling. When
these units are added to the model they are redistributed to the
appropriate test status by the simulator.
}
\author{
Thomas Rosendal
}
