% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{counti}
\alias{counti}
\title{counti}
\usage{
counti(
  model,
  class = c("true", "calf", "heifer", "cow", "apparent", "milk", "nmilk"),
  type = c("nop", "wnp")
)
}
\arguments{
\item{model}{A paratb model object}

\item{class}{The age category or sample type that you want the
prevalance from}

\item{type}{The type of prevalence}
}
\value{
A data.frame from the SimInf::prevalance method

A data.frame from the SimInf::prevalance method
}
\description{
A method to count the number in some compartments.
}
\details{
Similar to prev but without the denomenator
}
\author{
Thomas Rosendal
}
