% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{missed}
\alias{missed}
\title{missed}
\usage{
missed(model)
}
\arguments{
\item{model}{A paratb model object}
}
\value{
A data.frame
}
\description{
A method to count the number of herds missed by the surveillance
}
\author{
Thomas Rosendal
}
