% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{paratb}
\alias{paratb}
\title{Create a model for the SimInf framework}
\usage{
paratb(u0 = NULL, tspan = NULL)
}
\arguments{
\item{u0}{A data.frame with the initial state in each node.}

\item{tspan}{A vector (length >= 2) of increasing time points
where the state of each node is to be returned.}
}
\description{
Create a model to be used by the SimInf framework.
}
