% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{prev}
\alias{prev}
\title{prev}
\usage{
prev(
  model,
  class = c("true", "calf", "heifer", "cow", "apparent", "milk"),
  type = c("nop", "wnp"),
  ...
)
}
\arguments{
\item{model}{A paratb model object}

\item{class}{The age category or sample type that you want the
prevalance from}

\item{type}{the type pf prevalance}

\item{...}{Other arguments}
}
\value{
A data.frame from the SimInf::prevalance method
}
\description{
A method for calculating prevalance in a paratb model
}
\details{
This just assembles the pieces to pass onto the SimInf::prevalance
function.
}
\author{
Thomas Rosendal
}
