% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{prev_class}
\alias{prev_class}
\title{prevclass}
\usage{
prev_class(class = c("true", "calf", "heifer", "cow", "apparent", "milk"))
}
\arguments{
\item{class}{Age class of the prevalence}
}
\value{
A formula
}
\description{
prev_class
}
\author{
Thomas Rosendal
}
