% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{random_events}
\alias{random_events}
\title{random_events}
\usage{
random_events(model, n)
}
\arguments{
\item{model}{The model that the random event should be added to}

\item{n}{The number of random events}
}
\value{
A model object
}
\description{
random_events
}
\details{
A method to add random external transfer events to a model. This
is to illustrate the model working with events despite that we
cannot share the recording events used in the study.
}
