% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{seed_herd}
\alias{seed_herd}
\title{seed_herd}
\usage{
seed_herd(model, i, age = c("calf", "heifer", "cow"), n, seed = NULL)
}
\arguments{
\item{model}{The model object}

\item{i}{the node to seed in}

\item{age}{the age category to seed in. One of c("calf", "heifer", "cow").}

\item{n}{the number of animals to seed}

\item{seed}{The seed}
}
\value{
a paratb model object
}
\description{
seed_herd
}
\details{
A method for seeding infected animals in a herd.
}
\author{
Thomas Rosendal
}
