% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{select_type}
\alias{select_type}
\title{select_type}
\usage{
select_type(type = c("all", "dairy", "beef"))
}
\arguments{
\item{type}{the herdtype.}
}
\value{
a vector of herd IDs
}
\description{
A method to select the type of herd in the herdtype dataset
}
