% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{which_pos}
\alias{which_pos}
\title{which_pos}
\usage{
which_pos(model)
}
\arguments{
\item{model}{A paratb model object}
}
\value{
A data.frame
}
\description{
A method to list the positive herds for each tstep
}
\author{
Thomas Rosendal
}
