% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run,paratb-method}
\alias{run,paratb-method}
\title{Run the model}
\usage{
\S4method{run}{paratb}(model, solver = NULL)
}
\arguments{
\item{model}{The model to run.}

\item{solver}{Which numerical solver to utilize. Default is 'ssa'.}
}
\value{
model with result from simulation.
}
\description{
Run the model
}
